+++ src/QNEthernet.cpp
@@ -227,6 +227,7 @@ bool EthernetClass::start() {
 
   // Initialize Ethernet, set up the callback, and set the netif to UP
   netif_ = enet_netif();
+  driver_set_chip_select_pin(chipSelectPin_);
   if (!enet_init(mac_, &netifEventFunc)) {
     return false;
   }