+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to
 * Redesigned the driver interface.
 * Un-deprecated `EthernetClass::begin(mac, timeout)` and
   `begin(mac, ip, dns, gateway, subnet)`.
+* Un-deprecated `EthernetClass::init(sspin)` and added a driver function for
+  setting the chip select pin. The type of `sspin` was also changed to `int`.
 
 ### Fixed
 * Added missing `hostByName` to keywords.txt.